bump

+++ src/felix86/emulator.cpp
@@ -227,7 +227,7 @@ std::pair<ExitReason, int> Emulator::Start(const StartParameters& config) {
     ASSERT(path.find(g_config.rootfs_path.string()) == 0);
     g_params.argv[0] = path;
 
-    for (int i = 0; i < g_params.argv.size(); i++) {
+    for (size_t i = 0; i < g_params.argv.size(); i++) {
         // We need to remove any rootfs prefix from the arguments
         auto& p = g_params.argv[i];
         if (p.find(g_config.rootfs_path) == 0) {
